+++ backend/alembic/env.py
@@ -37,6 +37,7 @@ from app.models.datasource_table import DataSourceTable
 from app.models.git_repository import GitRepository
 from app.models.metadata_indexing_job import MetadataIndexingJob
 from app.models.dbt_resource import DBTResource
+from app.models.organization_settings import OrganizationSettings
 
 from app.settings.config import settings