+++ release_notes.md
@@ -8,6 +8,7 @@
 - Fixed memory bug
 - Added reasoning for planner agent
 - Added data preview for LLM to achieve ReAct like flow with code generation
+- Added organization settings to control AI features (specific agent skills) and additional settings (LLM viewing data, etc)
 
 ## Version 0.0.154 (March 24, 2025)